feat: implement the basic structure of cluster mode
