+++ project/rkl/src/commands/mod.rs
@@ -2,6 +2,7 @@ use clap::Parser;
 use std::fs;
 use std::io;
 use std::path::{Path, PathBuf};
+use std::str::FromStr;
 
 use anyhow::{Context, Result, anyhow, bail};
 use libcontainer::container::Container;
@@ -381,9 +382,9 @@ impl From<ExecContainer> for Exec {
 
 fn parse_env<T, U>(s: &str) -> Result<(T, U), Box<dyn Error + Send + Sync + 'static>>
 where
-    T: std::str::FromStr,
+    T: FromStr,
     T::Err: Error + Send + Sync + 'static,
-    U: std::str::FromStr,
+    U: FromStr,
     U::Err: Error + Send + Sync + 'static,
 {
     let pos = s
@@ -394,9 +395,9 @@ where
 
 fn parse_user<T, U>(s: &str) -> Result<(T, Option<U>), Box<dyn Error + Send + Sync + 'static>>
 where
-    T: std::str::FromStr,
+    T: FromStr,
     T::Err: Error + Send + Sync + 'static,
-    U: std::str::FromStr,
+    U: FromStr,
     U::Err: Error + Send + Sync + 'static,
 {
     if let Some(pos) = s.find(':') {
