+++ project/rkl/src/commands/pod/cluster.rs
@@ -1,0 +1,14 @@
+pub fn delete_pod(pod_name: &str) -> Result<(), anyhow::Error> {
+    println!("cluster delete the pod");
+    todo!()
+}
+
+pub fn create_pod(pod_yaml: &str) -> Result<(), anyhow::Error> {
+    println!("cluster create the pod");
+    todo!()
+}
+
+pub fn list_pod() -> Result<(), anyhow::Error> {
+    println!("cluster list the pod");
+    todo!()
+}
