+++ project/rkl/src/commands/pod/mod.rs
@@ -1,16 +1,19 @@
-use crate::commands::{Exec, ExecPod};
-use crate::commands::{delete, exec, load_container, start, state};
+use crate::commands::pod::standalone::{exec_pod, start_pod, state_pod};
 use crate::rootpath;
-use crate::task::{self, TaskRunner};
+use crate::task::TaskRunner;
 use crate::{PodCommand, daemon};
 use anyhow::{Result, anyhow};
 use daemonize::Daemonize;
-use liboci_cli::{Delete, Start, State};
+use std::env;
 use std::fs::{self, File};
-use std::io::{Read, Write};
+use std::io::Read;
+use std::io::Write;
 use std::path::{Path, PathBuf};
 use std::time::{SystemTime, UNIX_EPOCH};
-use tracing::{error, info};
+use tracing::info;
+
+pub mod cluster;
+pub mod standalone;
 
 // store infomation of pod
 #[derive(Debug)]
@@ -115,172 +118,6 @@ pub fn run_pod(pod_yaml: &str) -> Result<(), anyhow::Error> {
     run_pod_from_taskrunner(task_runner)
 }
 
-pub fn create_pod(pod_yaml: &str) -> Result<(), anyhow::Error> {
-    let mut task_runner = TaskRunner::from_file(pod_yaml)?;
-    let pod_name = task_runner.task.metadata.name.clone();
-
-    let pod_request = task_runner.build_run_pod_sandbox_request();
-    let config = pod_request
-        .config
-        .as_ref()
-        .ok_or_else(|| anyhow!("PodSandbox config is required"))?;
-    task_runner.sandbox_config = Some(config.clone());
-    let pod_response = task_runner.run_pod_sandbox(pod_request)?;
-    let pod_sandbox_id = pod_response.pod_sandbox_id;
-
-    let pause_pid = task_runner.pause_pid.ok_or_else(|| {
-        anyhow!(
-            "Pause container PID not found for PodSandbox ID: {}",
-            pod_sandbox_id
-        )
-    })?;
-    info!(
-        "PodSandbox (Pause) created: {}, pid: {}\n",
-        pod_sandbox_id, pause_pid
-    );
-
-    let mut container_ids = Vec::new();
-    for container in &task_runner.task.spec.containers {
-        let create_request =
-            task_runner.build_create_container_request(&pod_sandbox_id, container)?;
-        let create_response = task_runner.create_container(create_request)?;
-        container_ids.push(create_response.container_id.clone());
-        info!(
-            "Container created: {} (ID: {})",
-            container.name, create_response.container_id
-        );
-    }
-
-    let root_path = rootpath::determine(None)?;
-    let pod_info = PodInfo {
-        pod_sandbox_id,
-        container_names: container_ids,
-    };
-    pod_info.save(&root_path, &pod_name)?;
-
-    info!("Pod {} created successfully", pod_name);
-    Ok(())
-}
-
-pub fn start_pod(pod_name: &str) -> Result<(), anyhow::Error> {
-    let root_path = rootpath::determine(None)?;
-    let pod_info = PodInfo::load(&root_path, pod_name)?;
-
-    if pod_info.container_names.is_empty() {
-        return Err(anyhow!("No containers found for Pod {}", pod_name));
-    }
-
-    for container_name in &pod_info.container_names {
-        let start_args = Start {
-            container_id: container_name.clone(),
-        };
-        start(start_args, root_path.clone())
-            .map_err(|e| anyhow!("Failed to start container {}: {}", container_name, e))?;
-        info!("Container started: {}", container_name);
-    }
-
-    info!("Pod {} started successfully", pod_name);
-    Ok(())
-}
-
-pub fn delete_pod(pod_name: &str) -> Result<(), anyhow::Error> {
-    let root_path = rootpath::determine(None)?;
-    let pod_info = PodInfo::load(&root_path, pod_name)?;
-    let container = load_container(root_path.clone(), pod_name)
-        .map_err(|e| anyhow!("Failed to load container {}: {}", pod_name, e))?;
-    let pid_i32 = container
-        .state
-        .pid
-        .ok_or_else(|| anyhow!("PID not found for container {}", pod_name))?;
-    remove_pod_network(pid_i32)?;
-    // delete all container
-    for container_name in &pod_info.container_names {
-        let delete_args = Delete {
-            container_id: container_name.clone(),
-            force: true,
-        };
-        let root_path = rootpath::determine(None)?;
-        if let Err(delete_err) = delete(delete_args, root_path.clone()) {
-            error!(
-                "Failed to delete container {}: {}",
-                container_name, delete_err
-            );
-        } else {
-            info!("Container deleted: {}", container_name);
-        }
-    }
-
-    // delete pause container
-    let delete_args = Delete {
-        container_id: pod_info.pod_sandbox_id.clone(),
-        force: true,
-    };
-    let root_path = rootpath::determine(None)?;
-    if let Err(delete_err) = delete(delete_args, root_path.clone()) {
-        error!(
-            "Failed to delete PodSandbox {}: {}",
-            pod_info.pod_sandbox_id, delete_err
-        );
-    } else {
-        info!("PodSandbox deleted: {}", pod_info.pod_sandbox_id);
-    }
-
-    // delete pod file
-    PodInfo::delete(&root_path, pod_name)?;
-    info!("Pod {} deleted successfully", pod_name);
-    Ok(())
-}
-
-pub fn remove_pod_network(pid: i32) -> Result<(), anyhow::Error> {
-    let mut cni = task::get_cni()?;
-    cni.load_default_conf();
-
-    let netns_path = format!("/proc/{pid}/ns/net");
-    let id = pid.to_string();
-    cni.remove(id, netns_path.clone())
-        .map_err(|e| anyhow::anyhow!("Failed to remove CNI network: {}", e))?;
-
-    Ok(())
-}
-
-pub fn state_pod(pod_name: &str) -> Result<(), anyhow::Error> {
-    let root_path = rootpath::determine(None)?;
-    let pod_info = PodInfo::load(&root_path, pod_name)?;
-
-    println!("Pod: {pod_name}");
-
-    println!("PodSandbox ID: {}", pod_info.pod_sandbox_id);
-    let _ = state(
-        State {
-            container_id: pod_info.pod_sandbox_id.clone(),
-        },
-        root_path.clone(),
-    );
-
-    println!("Containers:");
-    for container_name in &pod_info.container_names {
-        let _container_state = state(
-            State {
-                container_id: container_name.clone(),
-            },
-            root_path.clone(),
-        );
-    }
-
-    Ok(())
-}
-
-pub fn exec_pod(args: ExecPod) -> Result<i32> {
-    let root_path = rootpath::determine(None)?;
-    let pod_info_path = root_path.join("pods").join(&args.pod_name);
-    if !pod_info_path.exists() {
-        return Err(anyhow::anyhow!("Pod {} not found", args.pod_name));
-    }
-    let args = Exec::from(args);
-    let exit_code = exec(args, root_path)?;
-    Ok(exit_code)
-}
-
 pub fn set_daemonize() -> Result<(), anyhow::Error> {
     let log_path = PathBuf::from("/var/log/rk8s/");
     if !log_path.exists() {
@@ -312,14 +149,39 @@ pub fn start_daemon() -> Result<(), anyhow::Error> {
 pub fn pod_execute(cmd: PodCommand) -> Result<()> {
     match cmd {
         PodCommand::Run { pod_yaml } => run_pod(&pod_yaml),
-        PodCommand::Create { pod_yaml } => create_pod(&pod_yaml),
+        PodCommand::Create { pod_yaml, cluster } => pod_create(&pod_yaml, cluster),
         PodCommand::Start { pod_name } => start_pod(&pod_name),
-        PodCommand::Delete { pod_name } => delete_pod(&pod_name),
+        PodCommand::Delete { pod_name, cluster } => pod_delete(&pod_name, cluster),
         PodCommand::State { pod_name } => state_pod(&pod_name),
         PodCommand::Exec(exec) => {
             let exit_code = exec_pod(*exec)?;
             std::process::exit(exit_code);
         }
         PodCommand::Daemon => start_daemon(),
+        PodCommand::List { .. } => pod_list(),
+    }
+}
+
+fn pod_list() -> Result<()> {
+    cluster::list_pod()
+}
+
+fn pod_delete(pod_name: &str, cluster: bool) -> Result<()> {
+    let is_cluster: bool = env::var("RKL_POD_CLUSTER")
+        .unwrap_or("false".to_string())
+        .parse()?;
+    match cluster || is_cluster {
+        false => standalone::delete_pod(pod_name),
+        true => cluster::delete_pod(pod_name),
+    }
+}
+
+fn pod_create(pod_yaml: &str, cluster: bool) -> Result<()> {
+    let is_cluster: bool = env::var("RKL_POD_CLUSTER")
+        .unwrap_or("false".to_string())
+        .parse()?;
+    match cluster || is_cluster {
+        false => standalone::create_pod(pod_yaml),
+        true => cluster::create_pod(pod_yaml),
     }
 }
