+++ project/rkl/tests/test_pod.rs
@@ -2,10 +2,10 @@ use crate::test_common::get_pod_config;
 use anyhow::anyhow;
 //use rkl::task::PodTask;
 use common::{ContainerRes, PodTask, Resource};
-use rkl::{
-    commands::pod::{create_pod, delete_pod, run_pod, start_pod},
-    task::TaskRunner,
-};
+use rkl::commands::pod::standalone::create_pod;
+use rkl::commands::pod::standalone::delete_pod;
+use rkl::commands::pod::standalone::start_pod;
+use rkl::{commands::pod::run_pod, task::TaskRunner};
 use serde_json::Value;
 use serial_test::serial;
 use std::{env, fs::File, io::Write, path::Path};
